+++ core/src/main/scala/sgl/geometry/Rect.scala
@@ -1,11 +1,11 @@
 package sgl.geometry
 
 //TODO: we should differential AABB and OBB
-case class Rect(left: Int, top: Int, width: Int, height: Int) {
+case class Rect(left: Float, top: Float, width: Float, height: Float) {
   require(width >= 0 && height >= 0)
 
-  def right: Int = left + width
-  def bottom: Int = top + height
+  def right: Float = left + width
+  def bottom: Float = top + height
 
   def centerX = left + width/2
   def centerY = top + height/2
@@ -22,10 +22,10 @@ case class Rect(left: Int, top: Int, width: Int, height: Int) {
   def vertices: Set[Point] = Set(topLeft, topRight, bottomLeft, bottomRight)
 
   //maybe intersect should go into external objects since there is no notion of direction (point vs rect)
-  def intersect(x: Int, y: Int): Boolean =
+  def intersect(x: Float, y: Float): Boolean =
     x >= left && x <= right && y >= top && y <= bottom
 
-  def intersect(point: Point): Boolean = intersect(point.x.toInt, point.y.toInt)
+  def intersect(point: Point): Boolean = intersect(point.x, point.y)
 
   def intersect(rect: Rect): Boolean = Collisions.aabbWithAabb(this, rect)
 
@@ -35,7 +35,7 @@ case class Rect(left: Int, top: Int, width: Int, height: Int) {
 
 object Rect {
 
-  def fromCoordinates(left: Int, top: Int, right: Int, bottom: Int): Rect =
+  def fromCoordinates(left: Float, top: Float, right: Float, bottom: Float): Rect =
     Rect(left, top, right - left, bottom - top)
 
 }
